feat(conanfile): add option with_wayland
Binaries created with this recipe can be different depending on the
distro version, and they will still have the same package hash(because
same settings and options). An improvement could be to add a with_wayland
option with values true,false and auto(the default value). In configure,
if with_wayland is auto, we set it to true or false depending on the distro.